+++ wezesha/src/App.js
@@ -1,14 +1,12 @@
 
-import Welcome from "./vision-mission"
+import Welcome from "./Vision-Mission"
 
 
 const wezeshaApp =()=>{
   return(
     <div>
       <h1></h1>
-     
       <Welcome/>
-      
     </div>
   )
 }
