+++ wezesha/src/Vision-Mission/index.css
@@ -1,0 +1,383 @@
+@import url('https://fonts.googleapis.com/css2?family=Karma:wght@300;400;500;600;700&display=swap');
+
+body{
+    display: flex;
+    margin-left: 4%;
+    font-size: 36px;
+    font-family: 'Karma',sans-serif;
+
+}
+
+.mission-text{
+    width:70%;
+    flex-wrap: wrap;
+    
+}
+
+.vision-text{
+    width:90%;
+    flex-wrap: wrap;
+    margin-left: -63px;
+   
+}
+
+#mission-image{
+    margin-left: 10%;
+    width: 650px;
+    height: 319;
+}
+#vision-image{
+    margin-left: -74px;
+    width: 518px;
+    height: 319;
+}
+
+@media screen and (min-width:280px) and (max-width:380px){
+    body{
+        display: block;
+        margin-left: -3.5%;
+        font-size: 20px;
+        margin-top: 5%;
+        font-family: 'Karma',sans-serif;
+        margin-bottom: 8%;
+    
+    }
+
+    .mission-text{
+        width:80%;
+        margin-left: 63px;
+        flex-wrap: wrap;
+        
+    }
+
+    .vision-text{
+        width:80%;
+        flex-wrap: wrap;
+        margin-left: 63px;
+       
+    }
+
+    #mission-image{
+        margin-left: 29%;
+        width: 217px;
+        height: 183px;
+    }
+
+    #vision-image{
+        margin-left: 25%;
+        width: 187px;
+        margin-top: 15px;
+
+    }
+
+}
+
+@media screen and (min-width:390px) and (max-width:414px){
+    body{
+        display: block;
+        margin-left: -3.5%;
+        font-size: 20px;
+        margin-top: 5%;
+        font-family: 'Karma',sans-serif;
+        margin-bottom: 8%;
+    
+    }
+
+    .mission-text{
+        width:80%;
+        margin-left: 63px;
+        flex-wrap: wrap;
+        
+    }
+
+    .vision-text{
+        width:80%;
+        flex-wrap: wrap;
+        margin-left: 60px;
+       
+    }
+
+    #mission-image{
+        margin-left: 25%;
+        width: 264px;
+        height: 213px;
+    }
+
+    #vision-image{
+        margin-left: 25%;
+        width: 210px;
+        height: 213px;
+        margin-top: 63px;
+    }
+
+}
+
+@media screen and (min-width:415px) and (max-width:450px){
+    body{
+        display: block;
+        margin-left: -3.5%;
+        font-size: 20px;
+        margin-top: 16%;
+        font-family: 'Karma',sans-serif;
+
+    
+    }
+
+    .mission-text{
+        width:80%;
+        margin-left: 63px;
+        flex-wrap: wrap;
+        
+    }
+
+    .vision-text{
+        width:80%;
+        flex-wrap: wrap;
+        margin-left: 60px;
+       
+    }
+
+    #mission-image{
+        margin-left: 23%;
+        width: 264px;
+        height: 213px;
+    }
+
+    #vision-image{
+        margin-left: 23%;
+        width: 210px;
+        height: 213px;
+        margin-top: 20%;
+    }
+
+}
+
+@media screen and (min-width:460px) and (max-width:560px){
+    body{
+        display: block;
+        margin-left: -3.5%;
+        font-size: 30px;
+        margin-top: 5%;
+        font-family: 'Karma',sans-serif;
+        margin-bottom: 10%;
+
+    
+    }
+
+    .mission-text{
+        width:120%;
+        margin-left: 100px;
+        flex-wrap: wrap;
+        
+    }
+
+    .vision-text{
+        width:120%;
+        flex-wrap: wrap;
+        margin-left: 100px;
+       
+    }
+
+    #mission-image{
+        margin-left: 43%;
+        width: 367px;
+    }
+
+    #vision-image{
+        margin-left: 43%;
+        width: 310px;
+    }
+
+}
+
+@media screen and (min-width:670px) and (max-width:780px){
+    body{
+        display: block;
+        margin-left: -3.5%;
+        font-size: 30px;
+        margin-top: 5%;
+        font-family: 'Karma',sans-serif;
+
+    
+    }
+
+    .mission-text{
+        width:80%;
+        margin-left: 115px;
+        flex-wrap: wrap;
+        
+    }
+
+    .vision-text{
+        width:80%;
+        flex-wrap: wrap;
+        margin-left: 125px;
+       
+    }
+
+    #mission-image{
+        margin-left: 40%;
+        width: 264px;
+        height: 213px;
+    }
+
+    #vision-image{
+        margin-left: 40%;
+        width: 210px;
+        height: 213px;
+    }
+
+}
+
+@media screen and (min-width:790px) and (max-width:890px){
+    body{
+        display: block;
+        margin-left: -3.5%;
+        font-size: 30px;
+        margin-top: 5%;
+        font-family: 'Karma',sans-serif;
+        margin-bottom: 33px;
+    
+    }
+
+    .mission-text{
+        width:80%;
+        margin-left: 160px;
+        flex-wrap: wrap;
+        
+    }
+
+    .vision-text{
+        width:80%;
+        flex-wrap: wrap;
+        margin-left: 155px;
+       
+    }
+
+    #mission-image{
+        margin-left: 33%;
+        width: 400px;
+    }
+
+    #vision-image{
+        margin-left: 35%;
+        width: 300px;
+        margin-top: 10%;
+    }
+
+}
+
+@media screen and (min-width:900px) and (max-width:1000px){
+    body{
+        display: block;
+        margin-left: -3.5%;
+        font-size: 40px;
+        margin-top: 5%;
+        font-family: 'Karma',sans-serif;
+        margin-bottom: 5%;
+
+    
+    }
+
+    .mission-text{
+        width:80%;
+        margin-left: 130px;
+        flex-wrap: wrap;
+        
+    }
+
+    .vision-text{
+        width:80%;
+        flex-wrap: wrap;
+        margin-left: 155px;
+       
+    }
+
+    #mission-image{
+        margin-left: 33%;
+        width: 420px;
+    }
+
+    #vision-image{
+        margin-left: 35%;
+        width: 360px;
+        margin-top: 10%;
+    }
+
+}
+
+@media screen and (min-width:1020px) and (max-width:1280px){
+    body{
+        display: block;
+        margin-left: -3.5%;
+        font-size: 20px;
+        margin-top: 5%;
+        font-family: 'Karma',sans-serif;
+        margin-bottom: 5%;
+    }
+
+    .mission-text{
+        width:60%;
+        margin-left: 25%;
+        flex-wrap: wrap;
+    }
+
+    .vision-text{
+        width:60%;
+        flex-wrap: wrap;
+        margin-left: 25%;
+    }
+
+    #mission-image{
+        margin-left: 40%;
+        width: 400px;
+    }
+
+    #vision-image{
+        margin-left: 40%;
+        width: 309px;
+        margin-top: 33px;
+    }
+
+}
+
+@media screen and (min-width:1280px) and (max-width:1400px){
+    body{
+        display: block;
+        margin-left: 17px;
+        font-size: 20px;
+        margin-top: 5%;
+        font-family: 'Karma',sans-serif;
+        margin-bottom: 5%;
+
+    
+    }
+
+    .mission-text{
+        width:50%;
+        margin-left: 25%;
+        flex-wrap: wrap;
+        
+    }
+
+    .vision-text{
+        width:50%;
+        flex-wrap: wrap;
+        margin-left: 255;
+       
+    }
+
+    #mission-image{
+        margin-left: 38%;
+        /* width: 23%;  */
+    }
+
+    #vision-image{
+        margin-left: 40%;
+        margin-top: 33px;
+        margin-left: 40%;
+        /* width: 18%; */
+    }
+
+}
+
