--- wezesha/src/vision-mission/index.css
@@ -1,322 +1,0 @@
-welcome{
-    background-color: #ffffff;
-    display: flex;
-    margin-left: -172px;
-}
-
-.our-mission{
-    color: rgb(0, 0, 0);
-    flex-wrap: wrap;
-    width: 30%;
-    margin-top: 27px;
-    margin-left: 58%;
-    font-size: 30px;
-    text-align: start;
-}
-.our-vission{
-    color: rgb(0, 0, 0);
-    width: 30%;
-    margin-top: 18px;
-    text-align: start;
-    margin-left: 23%;
-    flex-wrap: wrap;
-    font-size: 30px;
-}
-#image{
-    width: 323px;
-    height: 259px;
-    margin-top: -6px;
-    margin-left: 64%;
-}
-.imagee{
-    margin-top: -21.4%;
-    margin-left: -270px;
-}
-.mision{
-   margin-left: -57px;
-   margin-top: 2%;
-}
-#images{
-    width: 260px;
-    height: 25
-    5px;
-    margin-top: -39px;
-    margin-left: 30%; 
-}
- 
-@media screen and (min-width:280px) and (max-width:412px){
-        #image{
-            width: 190px;
-            height: 160px;
-            margin-top: 20px;
-            margin-left: 34%;
-        }
-    #images {
-        width: 150px;
-        height: 160px;
-        margin-top: 20px; 
-        margin-left: 54%; 
-        transform: translateX(-50%);
-    }
-    .welcome{
-        display: block;
-    }
-    .our-mission{
-        color: rgb(0, 0, 0);
-        flex-wrap: wrap;
-        width: 63%;
-        margin-top: 12px;
-        margin-left:100px;
-        font-size: 15px;
-        text-align: start;
-    }
-    .our-vission{
-        color: rgb(0, 0, 0);
-        width: 63%;
-        margin-top: 12px;
-        text-align: start;
-        margin-left: 120px;
-        flex-wrap: wrap;
-        font-size: 15px;
-    }
-    .imagee{
-        margin-top: 15%;
-        margin-left: -70px;
-    }
-    .mision{
-        margin-left: -57px;
-     }
-  }
-  @media screen and (min-width:414px) and (max-width:556px){
-    #image {
-        width: 208px;
-        height: 160px;
-        margin-top: 20px; 
-        margin-left: 59%; 
-        transform: translateX(-50%);
-    }
-
-    #images {
-        width: 170px;
-        height: 170px;
-        margin-top: 14px; 
-        margin-left: 58%; 
-        transform: translateX(-50%);
-    }
-    .welcome{
-        display: block;
-    }
-    .our-mission{
-        color: rgb(0, 0, 0);
-        flex-wrap: wrap;
-        width: 63%;
-        margin-top: 12px;
-        margin-left:140px;
-        font-size: 18px;
-        text-align: start;
-    }
-    .our-vission{
-        color: rgb(0, 0, 0);
-        width: 63%;
-        margin-top: 12px;
-        text-align: start;
-        margin-left: 150px;
-        flex-wrap: wrap;
-        font-size: 18px;
-    }
-    .imagee{
-        margin-top: 15%;
-        margin-left: -70px;
-    }
-    .mision{
-        margin-left: -57px;
-        margin-top: -6%;
-     }
-   
-  }
-  @media screen and (min-width:557px) and (max-width:739px){
-    #image {
-        width: 360px;
-        height: 250px;
-        margin-top: 20px; 
-        margin-left: 105%; 
-        transform: translateX(-50%);
-    }
-
-    #images {
-        width: 308px;
-        height: 277px;
-        margin-top: 20px; 
-        margin-left: 100%; 
-        transform: translateX(-50%);
-    }
-    .welcome{
-        display: block;
-    }
-    .our-mission{
-        color: rgb(0, 0, 0);
-        flex-wrap: wrap;
-        width: 110%;
-        margin-top: 12px;
-        margin-left:380px;
-        font-size: 35px;
-        text-align: start;
-    }
-    .our-vission{
-        color: rgb(0, 0, 0);
-        width: 110%;
-        margin-top: 12px;
-        text-align: start;
-        margin-left: 384px;
-        flex-wrap: wrap;
-        font-size: 35px;
-    }
-    .imagee{
-        margin-top: 15%;
-        margin-left: -286px;
-    }
-    .mision{
-        margin-left: -287px;
-     }
-    }
-  
-  @media screen and (min-width:740px) and (max-width:886px){
-    #image {
-        width: 270px;
-        height: 220px;
-        margin-top: 20px; 
-        margin-left: 50%; 
-        transform: translateX(-50%);
-    }
-
-    #images {
-        width: 230px;
-        height: 220px;
-        margin-top: 20px; 
-        margin-left: 50%; 
-        transform: translateX(-50%);
-    }
-    .welcome{
-        display: block;
-    }
-    .our-mission{
-        color: rgb(0, 0, 0);
-        flex-wrap: wrap;
-        width: 63%;
-        margin-top: 12px;
-        margin-left:180px;
-        font-size: 25px;
-        text-align: start;
-    }
-    .our-vission{
-        color: rgb(0, 0, 0);
-        width: 63%;
-        margin-top: 12px;
-        text-align: start;
-        margin-left: 180px;
-        flex-wrap: wrap;
-        font-size: 25px;
-    }
-    .imagee{
-        margin-top: 15%;
-        margin-left: -70px;
-    }
-    .mision{
-        margin-left: -57px;
-        margin-top: -23px;
-     }
-  }
-  
-   @media screen and (min-width:887px) and (max-width:1100px){
-    #image {
-        width: 370px;
-        height: 280px;
-        margin-top: 20px; 
-        margin-left: 54%; 
-        transform: translateX(-50%);
-    }
-
-    #images {
-        width: 300px;
-        height: 320px;
-        margin-top: 20px; 
-        margin-left: 48%; 
-        transform: translateX(-50%);
-    }
-    .welcome{
-        display: block;
-    }
-    .our-mission{
-        color: rgb(0, 0, 0);
-        flex-wrap: wrap;
-        width: 63%;
-        margin-top: 12px;
-        margin-left:180px;
-        font-size: 25px;
-        text-align: start;
-    }
-    .our-vission{
-        color: rgb(0, 0, 0);
-        width: 53%;
-        margin-top: 12px;
-        text-align: start;
-        margin-left: 200px;
-        flex-wrap: wrap;
-        font-size: 25px;
-    }
-    .imagee{
-        margin-top: 15%;
-        margin-left: 30px;
-    }
-    .mision{
-        margin-left: 30px;
-     }
-     
-  } 
-  
-  @media screen and (min-width:1101px) and (max-width:1300px){
-    #image {
-        width: 460px;
-        height: 350px;
-        margin-top: 20px; 
-        margin-left: 53%; 
-        transform: translateX(-50%);
-    }
-
-    #images {
-        width: 370px;
-        height: 360px;
-        margin-top: 20px; 
-        margin-left: 43%; 
-        transform: translateX(-50%);
-    }
-    .welcome{
-        display: block;
-    }
-    .our-mission{
-        color: rgb(0, 0, 0);
-        flex-wrap: wrap;
-        width: 63%;
-        margin-top: 12px;
-        margin-left:180px;
-        font-size:35px;
-        text-align: start;
-    }
-    .our-vission{
-        color: rgb(0, 0, 0);
-        width: 53%;
-        margin-top: 12px;
-        text-align: start;
-        margin-left: 200px;
-        flex-wrap: wrap;
-        font-size: 35px;
-    }
-    .imagee{
-        margin-top: 15%;
-        margin-left: 50px;
-    }
-    .mision{
-        margin-left: 37px;
-     }
-   
-  }
